% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ECI_data}
\alias{get_ECI_data}
\title{Access data sets used for tutorials and problem sets}
\usage{
get_ECI_data(name = NULL)
}
\arguments{
\item{name}{The name of the data set}
}
\value{
An object, typically a data frame
}
\description{
Access data sets used for tutorials and problem sets
}
\examples{
satisfaction <- get_ECI_data('problem-set-2-satisfaction')
}
